requirements.txt
Django
psycopg2
django-bootstrap3
django-markdown
Pillow
django-imagekit
pytz
django-taggit
Markdown
django-haystack
pysolr
django-extensions
werkzeug
django-sitetree
django-markdownx
django-meta